requirements.txt
requests==2.31.0
python-telegram-bot==21.0.1
schedule==1.2.0
python-dotenv==1.0.0
pytz==2024.1
PyJWT==2.8.0
cryptography==41.0.7